refactor: move minimal testing content
