+++ tests/testcore/minimal_testing_content.cpp
