+++ tests/testcore/validation/character_race/character_race_data_test.cpp
@@ -8,7 +8,7 @@
 
 #include <core/errors/errors.hpp>
 #include <core/validation/feature/feature_data.hpp>
-#include <testcore/validation/minimal_testing_content.hpp>
+#include <testcore/minimal_testing_content.hpp>
 #include <testcore/validation/validation_data_mock.hpp>
 
 static constexpr const char* tags = "[core][validation][character_race]";
