+++ tests/testcore/validation/effect_holder/subholders/riv_holder_data_test.cpp
@@ -5,7 +5,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 #include <core/errors/errors.hpp>
-#include <testcore/validation/minimal_testing_content.hpp>
+#include <testcore/minimal_testing_content.hpp>
 #include <testcore/validation/validation_data_mock.hpp>
 
 static constexpr const char* tags = "[core][validation][effect_holder]";
